[DF-415] Moved GL code to Texture class

+++ Sources/Internal/Render/TextureDescriptor.h
@@ -132,8 +132,8 @@ public:
     
     int8 generateMipMaps;
     
-    int32 minFilter;
-    int32 magFilter;
+    int8 minFilter;
+    int8 magFilter;
     
     
     Compression pvrCompression;
